modify setFlushInterval arg type int -> long

+++ rake/src/test/java/com/skp/di/rake/client/utils/RakeTestUtils.java
@@ -114,7 +114,7 @@ public class RakeTestUtils {
             }
 
             @Override
-            public int provideFlushIntervalAsMilliseconds() {
+            public long provideFlushIntervalAsMilliseconds() {
                 return intervalAsMilliseconds;
             }
